requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
telebot==0.0.5
python-dotenv==1.0.1